// src/data/sanitizedLayerStats.generated.ts
// GENERATED FILE - DO NOT EDIT MANUALLY
export const sanitizedLayerStats = {
  generatedAt: "2026-06-10T20:25:05.380Z",
  layers: [
  {
    "id": "conectores_visualizacion",
    "path": "/data/sanitized/conectores_visualizacion.public.geojson",
    "exists": true,
    "sizeMB": 0,
    "featureCount": 5,
    "geometryTypes": [
      "LineString"
    ],
    "bbox": [
      -98.89283437207115,
      19.490583058746367,
      -98.88300015429346,
      19.496304051408757
    ],
    "renderable": true
  },
  {
    "id": "evidencia_documental_agregada_por_zona",
    "path": "/data/sanitized/evidencia_documental_agregada_por_zona.public.geojson",
    "exists": true,
    "sizeMB": 0.03,
    "featureCount": 18,
    "geometryTypes": [
      "MultiPolygon"
    ],
    "bbox": [
      -98.93109111098386,
      19.48360196092422,
      -98.87423175715932,
      19.50551611973924
    ],
    "renderable": true
  },
  {
    "id": "nodos_documentales_publicos_agregados",
    "path": "/data/sanitized/nodos_documentales_publicos_agregados.public.geojson",
    "exists": true,
    "sizeMB": 0.02,
    "featureCount": 20,
    "geometryTypes": [
      "Point"
    ],
    "bbox": [
      -98.9106280898513,
      19.487188729788194,
      -98.88227626819058,
      19.50021316314282
    ],
    "renderable": true
  },
  {
    "id": "nodos_orientacion_base",
    "path": "/data/sanitized/nodos_orientacion_base.public.geojson",
    "exists": true,
    "sizeMB": 0.04,
    "featureCount": 60,
    "geometryTypes": [
      "Point"
    ],
    "bbox": [
      -98.89401756564911,
      19.487924821324032,
      -98.87412325456951,
      19.496277974420053
    ],
    "renderable": true
  },
  {
    "id": "vias_contexto_movilidad",
    "path": "/data/sanitized/vias_contexto_movilidad.light.geojson",
    "exists": true,
    "sizeMB": 0.72,
    "featureCount": 1049,
    "geometryTypes": [
      "LineString"
    ],
    "bbox": [
      -98.97054314399998,
      19.459646855000074,
      -98.83498932299995,
      19.584680879000075
    ],
    "renderable": true
  },
  {
    "id": "zonas_prioridad_integrada_o_indicadores",
    "path": "/data/sanitized/zonas_prioridad_integrada_o_indicadores.public.geojson",
    "exists": true,
    "sizeMB": 0.01,
    "featureCount": 5,
    "geometryTypes": [
      "MultiPolygon"
    ],
    "bbox": [
      -98.93109111098386,
      19.48360196092422,
      -98.87423175715932,
      19.50551611973924
    ],
    "renderable": true
  }
]
};